+++ internals-js/src/__tests__/operations.test.ts
@@ -1,10 +1,11 @@
 import {
+    CompositeType,
   defaultRootName,
   Schema,
   SchemaRootKind,
 } from '../../dist/definitions';
 import { buildSchema } from '../../dist/buildSchema';
-import { MutableSelectionSet, Operation, operationFromDocument, parseOperation } from '../../dist/operations';
+import { MutableSelectionSet, Operation, operationFromDocument, parseOperation, parseSelectionSet } from '../../dist/operations';
 import './matchers';
 import { DocumentNode, FieldNode, GraphQLError, Kind, OperationDefinitionNode, OperationTypeNode, SelectionNode, SelectionSetNode } from 'graphql';
 
@@ -33,7 +34,7 @@ function astSSet(...selections: SelectionNode[]): SelectionSetNode {
 
 describe('fragments optimization', () => {
   // Takes a query with fragments as inputs, expand all those fragments, and ensures that all the
-  // fragments gets optimized back, and that we get back the exact same query. 
+  // fragments gets optimized back, and that we get back the exact same query.
   function testFragmentsRoundtrip({
     schema,
     query,
@@ -56,7 +57,7 @@ describe('fragments optimization', () => {
     // this are just about testing the reuse of fragments and this make things shorter/easier to write.
     // There is tests in `buildPlan.test.ts` that double-check that we don't reuse fragments used only
     // once in actual query plans.
-    const optimized = withoutFragments.optimize(operation.selectionSet.fragments!, 1);
+    const optimized = withoutFragments.optimize(operation.selectionSet.fragments!, {minUsagesToOptimize: 1});
     expect(optimized.toString()).toMatchString(operation.toString());
   }
 
@@ -70,13 +71,13 @@ describe('fragments optimization', () => {
         b: Int
       }
 
-      type T1 {
+      type T1 implements I {
         a: Int
         b: Int
         u: U
       }
 
-      type T2 {
+      type T2 implements I {
         x: String
         y: String
         b: Int
@@ -1263,4 +1264,238 @@ describe('unsatisfiable branches removal', () => {
   ])('removes unsatisfiable branches', ({input, output}) => {
     expect(withoutUnsatisfiableBranches(input)).toBe(output);
   });
+});
+
+test('contains ignores unecessary fragments even when subtyping is involved', () => {
+  const schema = parseSchema(`
+      type Query {
+        a: A!
+      }
+
+      interface IA1 {
+        b: IB1!
+      }
+
+      interface IA2 {
+        b: IB2!
+      }
+
+      type A implements IA1 & IA2 {
+        b: B!
+      }
+
+      interface IB1 {
+        v1: Int!
+      }
+
+      interface IB2 {
+        v2: Int!
+      }
+
+      type B implements IB1 & IB2 {
+        v1: Int!
+        v2: Int!
+      }
+  `);
+
+  const typeA = schema.type('A') as CompositeType;
+
+  const s1 = parseSelectionSet({
+    parentType: typeA,
+    source: '{ b { v1 v2 } }'
+  });
+
+  const s2 = parseSelectionSet({
+    parentType: typeA,
+    source: '{ ... on IA1 { b { v1 } } ... on IA2 { b { v2 } } }'
+  });
+
+  // Here, A is a concrete type, and IA1 and IA2 are just 2 of its interfaces, so
+  //   a { ... on IA1 { b { v1 } } ... on IA2 { b { v2 } } }
+  // is basically exactly the same as:
+  //   a { b { v1 } b { v2 } }
+  // which is the same as:
+  //   a { b { v1 v2 } }
+  // and that is why we want the `contains` below to work (note that a simple `contains`
+  // that doesn't handle this kind of subtlety could have its use too, it would just be
+  // a different contract, but we want "our" `contains` to handle this because of named
+  // fragments "reconstruction" where that kind of subtlety arises).
+  //
+  // Here, the added subtlety is that there is interface subtyping involved too.
+  expect(s2.contains(s1)).toBeTruthy();
+});
+
+test('Reuse fragment if possible else auto re-fragment', () => {
+  const schema = parseSchema(`
+    type Query {
+      t: T1
+    }
+    interface I {
+      b: Int
+    }
+    type T1 {
+      a: Int
+      b: Int
+      c: String
+      d: T2
+    }
+    type T2 {
+      b: Int
+      x: String
+      y: String
+      u: String
+      t: String
+      s: String
+      r: String
+      q: String
+      p: String
+      o: String
+      z: Int
+      v: String
+      w: T1
+    }
+    union U = T1 | T2
+  `);
+
+  const operation = parseOperation(schema, `
+    fragment OnSubT2 on T2 {
+      x
+      y
+      u
+      t
+      s
+      r
+      q
+      p
+    }
+    query {
+      t {
+        d {
+          ...OnSubT2
+          z
+          w {
+            d {
+              ...OnSubT2
+              v
+            }
+          }
+        }
+      }
+      duplicate: t {
+        d {
+          ...OnSubT2
+          z
+          w {
+            d {
+              ...OnSubT2
+              v
+            }
+          }
+        }
+      }
+    }
+  `);
+  const expandedOperationForQP = operation.expandAllFragments();
+  const resultantOperationWithoutFragments = expandedOperationForQP.optimize(operation.selectionSet.fragments!, {minUsagesToOptimize: 2,  autoFragmentize: false});
+  expect(resultantOperationWithoutFragments.toString()).toMatchString(`
+  fragment OnSubT2 on T2 {
+    x
+    y
+    u
+    t
+    s
+    r
+    q
+    p
+  }
+
+  {
+    t {
+      d {
+        ...OnSubT2
+        z
+        w {
+          d {
+            x
+            y
+            u
+            t
+            s
+            r
+            q
+            p
+            v
+          }
+        }
+      }
+    }
+    duplicate: t {
+      d {
+        ...OnSubT2
+        z
+        w {
+          d {
+            x
+            y
+            u
+            t
+            s
+            r
+            q
+            p
+            v
+          }
+        }
+      }
+    }
+  }`);
+  const optimized = expandedOperationForQP.optimize(operation.selectionSet.fragments!, {minUsagesToOptimize:2, autoFragmentize: true});
+  // Note that since auto refragment is set to true we were able to fragment the composite types
+  // and thus reduce the number of lines
+  // even if we could not re use the query fragment "OnSubT2"
+  expect(optimized.toString()).toMatchString(`
+  fragment OnSubT2 on T2 {
+    x
+    y
+    u
+    t
+    s
+    r
+    q
+    p
+  }
+
+  fragment T206a5e7e420bda1858b8f4a7b50df310b2444b2df0c2539a574eb73b2a47e331d on T2 {
+    ...OnSubT2
+    v
+  }
+
+  fragment T1ddac968f776e51fc51b44e4da419b068b627ba915c8470ef0db717b336da59a6 on T1 {
+    d {
+      ...T206a5e7e420bda1858b8f4a7b50df310b2444b2df0c2539a574eb73b2a47e331d
+    }
+  }
+
+  fragment T2150315cdc6028cf38902b0acb83322a8d03258910425d875326fb36072957f7f on T2 {
+    ...OnSubT2
+    z
+    w {
+      ...T1ddac968f776e51fc51b44e4da419b068b627ba915c8470ef0db717b336da59a6
+    }
+  }
+
+  fragment T1a064aab381a5268898c38068088deb7b1a4adf85e0b45762c82828c38df60878 on T1 {
+    d {
+      ...T2150315cdc6028cf38902b0acb83322a8d03258910425d875326fb36072957f7f
+    }
+  }
+
+  {
+    t {
+      ...T1a064aab381a5268898c38068088deb7b1a4adf85e0b45762c82828c38df60878
+    }
+    duplicate: t {
+      ...T1a064aab381a5268898c38068088deb7b1a4adf85e0b45762c82828c38df60878
+    }
+  }`);
 });